Make SqlClient sni.dll independent
