+++ src/System.Data.SqlClient/src/Interop/SNINativeMethodWrapper.Unix.cs
@@ -4,25 +4,6 @@
 
 using System.Data.SqlClient.SNI;
 
-namespace System.Data.SqlClient
-{
-    internal static class SNINativeMethodWrapper
-    {
-        internal enum SniSpecialErrors : uint
-        {
-            LocalDBErrorCode = SNICommon.LocalDBErrorCode,
-
-            // multi-subnet-failover specific error codes
-            MultiSubnetFailoverWithMoreThan64IPs = SNICommon.MultiSubnetFailoverWithMoreThan64IPs,
-            MultiSubnetFailoverWithInstanceSpecified = SNICommon.MultiSubnetFailoverWithInstanceSpecified,
-            MultiSubnetFailoverWithNonTcpProtocol = SNICommon.MultiSubnetFailoverWithNonTcpProtocol,
-
-            // max error code value
-            MaxErrorValue = SNICommon.MaxErrorValue
-        }
-    }
-}
-
 namespace System.Data
 {
     internal static partial class SafeNativeMethods
