+++ src/System.Data.SqlClient/src/System/Data/ProviderBase/DbConnectionPoolIdentity.Windows.cs
@@ -2,11 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-
-//------------------------------------------------------------------------------
-
+using System.Data.SqlClient;
 using System.Security.Principal;
-
 
 namespace System.Data.ProviderBase
 {
@@ -15,6 +12,11 @@ namespace System.Data.ProviderBase
         private static DbConnectionPoolIdentity s_lastIdentity = null;
 
         internal static DbConnectionPoolIdentity GetCurrent()
+        {
+            return TdsParserStateObjectFactory.UseManagedSNI ? GetCurrentManaged() : GetCurrentNative();
+        }
+
+        private static DbConnectionPoolIdentity GetCurrentNative()
         {
             DbConnectionPoolIdentity current;
             using (WindowsIdentity identity = WindowsIdentity.GetCurrent())
