+++ src/System.Data.SqlClient/src/System/Data/ProviderBase/DbConnectionPoolIdentity.cs
@@ -2,12 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-
-//------------------------------------------------------------------------------
-
-using System.Security.Principal;
-
-
 namespace System.Data.ProviderBase
 {
     sealed internal partial class DbConnectionPoolIdentity
@@ -47,6 +41,15 @@ namespace System.Data.ProviderBase
         override public int GetHashCode()
         {
             return _hashCode;
+        }
+
+        internal static DbConnectionPoolIdentity GetCurrentManaged()
+        {
+            string sidString = (!string.IsNullOrWhiteSpace(System.Environment.UserDomainName) ? System.Environment.UserDomainName + "\\" : "")
+                                + System.Environment.UserName;
+            bool isNetwork = false;
+            bool isRestricted = false;
+            return new DbConnectionPoolIdentity(sidString, isRestricted, isNetwork);
         }
     }
 }
